Replace double by BigDecimal

+++ src/main/java/com/genezeiniss/pos_transaction_processor/configuration/PaymentMethodProperties.java
@@ -4,10 +4,12 @@ import com.genezeiniss.pos_transaction_processor.domain.PriceModifierRange;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public abstract class PaymentMethodProperties {
 
-    private double pointsMultiplier;
+    private BigDecimal pointsMultiplier;
     private PriceModifierRange priceModifierRange;
 }
