+++ src/main/java/com/genezeiniss/pos_transaction_processor/domain/PriceModifierRange.java
@@ -1,4 +1,6 @@
 package com.genezeiniss.pos_transaction_processor.domain;
 
-public record PriceModifierRange(double min, double max) {
+import java.math.BigDecimal;
+
+public record PriceModifierRange(BigDecimal min, BigDecimal max) {
 }
