+++ src/main/java/com/genezeiniss/pos_transaction_processor/domain/Transaction.java
@@ -3,6 +3,7 @@ package com.genezeiniss.pos_transaction_processor.domain;
 import com.genezeiniss.pos_transaction_processor.domain.enums.PaymentMethod;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Map;
 
@@ -11,13 +12,11 @@ public class Transaction {
 
     protected String userId;
     protected String customerId;
-    protected double price;
-    protected double priceModifier;
+    protected BigDecimal price;
+    protected BigDecimal priceModifier;
     protected PaymentMethod paymentMethod;
     protected Instant datetime;
     protected Map<String, String> additionalInformation;
-    //todo: make sure that Graphql return values with two decimals
-    private double finalPrice;
+    private BigDecimal finalPrice;
     private int points;
-
 }
